Fix line endings in html files
The files were checked in with the line extension `CRLF`.
In the `.gitattributes` file, HTML files were defined so that they
should always have the native line endings of the operating system.
However, if the files are checked in with the line extension `CRLF`,
Git will not convert the line endings to native line endings at
checkout.

Related issues: #13