Add `G::CodeArchivPath`
Now `mod_G` stores the absolute path to the CodeArchiv root; it does so by
assuming that all modules and main codes that might use the modules reside
in the `_assets` folder (and thus that Archiv root is up one level).
Now `mod_CodeArchiv` preserves and restores the Current Directory of the tool
that invokes it, and relies on new `G::CodeArchivPath` to find the project's
root.