Make mod_G Abort if Can't Define Archiv Root
Add sanitation checks that `G::CodeArchivPath` actually point to the Archiv's
root, by checking for presence of `_assets` folder; if fails warns and aborts.
This is a safeguard to handle when the module is included by code not located
in the default paths, in which case it's supposed to manually set Curr Dir to
the root of the Archiv.